# public/swagger.json
{
 "swagger": "2.0",
 "info": {
  "version": "1.0.0",
  "title": "Backend API"
 },
 "host": "localhost:9000",
 "basePath": "/",
 "schemes": [
  "http"
 ],
 "consumes": [
  "application/json"
 ],
 "produces": [
  "application/json"
 ],
 "paths": {
...
  "/api/user/forgot-password": {
   "post": {
    "summary": "Forgot pasword",
    "description": "Forgot password it is and endpoint that send an email to user for recover his password",
    "operationId": "forgotPassword",
    "parameters": [
     {
      "name": "email",
      "in": "formData",
      "description": "The user email",
      "required": true,
      "type": "string"
     }
    ],
    "responses": {
     "200": {
      "description": "Send email with token for recover password successful!"
     },
     "404": {
      "description": "Route not found"
     },
     "500": {
      "description": "Something went wrong"
     }
    }
   }
  },
...
 }
}